+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 2.3.3
+- TBC: Warrior stances now show up as auras on sArena like normal (workaround required for TBC since they not *real* auras)
 - Localization added. Currently supports English and Korean. Thank you to 007bb for contributing.
 
 2.3.2b